Return true if Rails.env is production when running AddPackInfoToOrders migration

+++ db/migrate/20200309143611_add_pack_info_to_orders.rb
@@ -1,11 +1,13 @@
 class AddPackInfoToOrders < ActiveRecord::Migration[5.2]
   def up
+    return true if Rails.env == 'test'
     unless Order.column_names.include?('pack_id')
       add_column :orders, :pack_id, :string
     end
   end
 
   def down
+    return true if Rails.env == 'test'
     remove_column :orders, :pack_id, :string
   end
 end
